requirements.txt
certifi==2017.11.5
chardet==3.0.4
dj-database-url==0.4.2
Django==2.0.1
future==0.16.0
gunicorn==19.7.1
idna==2.6
line-bot-sdk==1.5.0
numpy==1.14.0
pandas==0.22.0
psycopg2==2.7.3.2
python-dateutil==2.6.1
pytz==2017.3
requests==2.18.4
six==1.11.0
urllib3==1.22
whitenoise==3.3.1
